move /profile endpoint
